Fix fullscreen and landscape task insets
Previously, we were incorrectly using the thumbnail insets, which
were insufficient for fullscreen tasks, and ignoring rotated tasks
entirely. Now we check what part of the thumbnail is actually
clipped (e.g. based on the scale we apply to get it to fit in the
same width as all the other tasks), and fill that out when setting
the fullscreen progress on each task.

Test:
- Quick switch (from home and from apps) to:
  - "Normal" apps such as Calculator in portrait and landscape
  - Immersive portrait apps such as Elder Scrolls: Blades
  - Immersive landscape apps such as Stardew Valley
- Swipe up from an app with the above cases in the adjacent task,
  ensure the thumbnail gets clipped from fullscreen to its final
  position in recents
- Do the above tests from landscape launcher as well

Bug: 130020567
Change-Id: I6f6b7f0ee134d464b3704990db9e1d3a01e6de0b