Fix bug where FolderIcon stays visible during swipe up to home animation.
This can happen when opening an app from a folder, and then immediately
swiping up back to home.

This happens because when we swipe up to go home, there's a race condition
where the folder close animation finishes and sets the FolderIcon to VISIBLE
right before FloatingIconView sets it to INVISIBLE.

To fix it, in OverviewState#onStateEnabled we call
AbstractFloatingView#closeAllOpenViews(animate=false). Then we added logic
to cancel any Folder closing animation (which there is,
from WindowTransformSwipeHandler#onLauncherStart) and to just close the
folder right then and there.

Bug: 132588097
Change-Id: I4379431815e7cbddede5ea0213fe9323f001484b