Improving swipe up interaction when device is locked
When device is locked, only scale down the top task as a response to
the user interaction. When user flings or lifts his finger, the task
is dismissed to go to the lock screen

LockScreenRecentsActivity is an empty activity which starts on top of lock
screen and finishes immediately. This allows us to start a recents transition
with just the top activity as the animation target. This target is then used
for swipe up interaction

Bug: 133167096
Change-Id: I466ed142ea33d626c78cb9cc5f6311bad26b8d98