Fix bug where changing the display size does not update the launcher grid.
The bug is caused by launcher saving the grid name and using that grid name
to look for matching display options. This makes sense when changing the grid
size, but doesn't work well when changing the display size.

Example: Initial Pixel display size is set to Default, so we save "normal" as
the KEY_IDP_GRID_NAME. When we change display size to Largest, the
KEY_IDP_GRID_NAME is still "normal" and so we only look at display options
under "normal". Before this, Pixel with display size set to Largest would be
set to "reasonable".

This should be safe change for Q, and we can have a proper fix when we
officially support changing grid size.

Bug: 131867841
Change-Id: If5f3b0a13b90069973e929024b26bd9b9c45a7d8