Android.mk: Strictly order QuickStepGo and GoIconRecents
We had Launcher3QuickStepGo adding Launcher3GoIconRecents to
LOCAL_OVERRIDES_PACKAGES, and also had Launcher3GoIconRecents
adding Launcher3QuickStepGo to LOCAL_OVERRIDES_PACKAGES.  As
a result, if both of these got added to PROJECT_PACKAGES, we
would end up with neither launcher, since they both overrode
each other.

To fix this, we declare a strict hierarchy.  Since
Launcher3GoIconRecents is the default, we put Launcher3QuickStepGo
to be higher in the hierarchy.  While that might seem a
counter-intuitive choice, it makes it much easier for folks
to override the default, without needing to modify the file
containing the default.

Test: Various clean builds which revealed this root issue, and this fix.
Bug: 135161289
Change-Id: If718a7030aa6ee4b6b952406006abda77b1eb8e5