Distinguish mDragHandleProgress from mMidProgress
mMidProgress is intended to be the point at which we reach Overview when
swiping up and thus start scrimming it as All Apps shows above it. For
that reason, mMidProgress needs to match OverviewState's progress. In
contrast, the drag handle (accessibility arrow) can move up sooner to
avoid overlapping with the QSB; for this, we added mDragHandleProgress.

Bug: 136800511
Change-Id: I2825ff798707530420855257350e492b4b9b77dc