Fix QSB becoming invisible
This could not be reproduced until I removed a line that wouldn't call
onAssistantVisiblityChanged if the argument was the same value as the
argument as the previous call.

After that, the bug became readily reproducible. I traced through
Launcher till I found that FallbackActivityControllerHelper
.onAssistantVisibilityChanged was being called while the screen was
locked, proving that onAssistantVisiblityChanged was _not_ reaching
launcher.

Test: Verify bug no longer reproduces.
Bug: 134981174
Bug: 135247753
Bug: 135572849
Bug: 135733393
Bug: 136386749
Bug: 136776987
Bug: 137534772
Bug: 137764419

Change-Id: Ib5e8df3b5030a77c5df351a1fcd993db6bd602fc