Set exclusion rect for launcher before it gains focus
Back gesture still shows if user attempts to swipe
back while we're animating to launcher state normal.
Focus is only granted after the animation has
completed, but the user can interact w/ the back
edge panels while it's ongoing.

fixes: 138622418
Test: Open app, swipe home and quickly swipe from
the edge of the screen. Should go to -1 screen and
not see back arrow.

Change-Id: I32228370c7ec1bdb75474fdff2d1c99cb677fa6a