Reduce risk of flakes in memory leak detection tests.
Now performing jreck@-style garbage collection prior to getting Launcher
PSS. The reduces PSS noise from ~700KB to ~250KB (std dev).

The tests (MemoryTests) run outside of Launcher process and
communicate with Launcher via RPC.

The change is in the RPC handler that lives in Launcher and returns
Launcher PSS to the test.

Bug: 139137636
Change-Id: I457e590b005dac53345f4223d972c7961754938c