When scrolling in All apps, scroll to last clickable row
Not the last row. This fixes tests broken after some assumption broke in
platform at an unknown moment.

Since the gesture distance became smaller in some cases, the speed of
the gesture became slower, and on some devices, the scroll gesture
triggers icon dragging.

To prevent this, I reduced the number of steps in the gesture to 10, but
added a slow-down at the end of the gesture to avoid inertia.

Also made some names and calculations clearer.

Bug: 144180777
Change-Id: Ie6c5776606ecff64d553fa836bdb3d90f32c5d9e