Align badging logic with platform IconDrawableFactory.
The current logic in BaseIconFactory is only trying to add the user
badge if the user passed in isn't null and isn't the current user,
then if no such attempt is made, try to add an instant app badge. This
is causing user badge to disappear if the current user is a work
profile. Meanwhile, the badging logic in IconDrawableFactory is to try
to add the instant app badge first, and then always try to add the
user badge.

Bug: 132625654
Test: manually test as in b/128625591

Change-Id: I2241bdb524ca6bc66e545a0ec8433119e0448358
Signed-off-by: Hyunyoung Song <hyunyoungs@google.com>