Fix search bar bottom margin in widget mode
The minimum bottom margin for the search bar is using device profile
insets intended for the hotseat search bar which are incorrect for
widgets. As a result, the widget is too high on the home screen and may
be cropped on some devices, especially if using more (and therefore
smaller) home screen rows than the default.

Fix by referencing the launcher's cell size relative to the widget
height rather than the device insets, making the search widget properly
centered in its row.

The calculation for non-widget mode is unchanged, but clean up some
duplicate calls to launcher.getDeviceProfile and getResources.

Fixes: https://github.com/LawnchairLauncher/Lawnchair/issues/1860
Signed-off-by: Allen Wild <allenwild93@gmail.com>