Use activity rotation to compute delta
When computing applying the rotation matrix to the TaskView thumnail,
use the rotation of the window configuration instead of the displays.

The forced rotation (ag/10176951) does not yet provide overrided values
for the display orientation when the requesting app is being forced
rotated, which is the case of the launcher. So
getDisplay().getRotation() would return the true rotation of the display
and not the rotation of the activity.

Test: with "adb shell settings put global forced_rotation 1"
and ag/10176951 and parents applied
 1. Open landscape application
 2. Display recents view
 3. Check that the thumnail is in the same rotation as the opened
 application

+ tests from I060a5dfcbcbb2ac21624d2ac778f76562bd60f40
Bug: 143053092
Bug: 143892437

Change-Id: Iee250459050b3c75dc3b79deed0682f809c7168b