Disables Back gesture while sandbox is running.
We'll want to detect this gesture ourselves to distinguish
between left and right edge and update the UI accordingly.

Note: This exclusion only works if the sandbox is launched
into the same task as Launcher (i.e. from Launcher itself).
I haven't found a way to get this to happen over adb.

Test: Launched BackGestureTutorialActivity from Launcher
and verified Back gesture animations did not play and the
onBackPressed callback was not called. Verified these did
still work in other apps.

Bug: 148542211
Change-Id: Iebc3428738edbf8ebfcff157f5952b787eb46bac