[DO NOT MERGE] Prevent hotseat touch hijack
Currently in AllAppsContinerView#onInterceptTouchEvent we check if a
touch event is in bounds of RecyclerViewFastScroller and set
mTouchHandler so the scroll bar handles onTouchEvent. However, if a user
performs a back gesture from the right side (which overlaps with the
scroller) and returns home, touches to the hotseat get intercepted.

Bug: 148639821
Test: Manual
Change-Id: Ie44f3a16b628b12ad00a7ae6c6bc5703171fbb56