Show IME suggestions while tapping on Folder EditText
Current behaviour doesn't show the suggestions in IME if the folder
name is non-empty (Shown only if the folder name is empty).

This change shows the IME suggestions always, but primary suggestions
are shown only if the folder name is empty (not overwritting the
current folder name with suggestion).

Bug: 151762006

Tested: on phone. (created a folder name, assigned custom name, added
    the third app to Folder, tap on Folder EditText and it shows IME
    suggestions)
Change-Id: Ic4d43660a371b8dd7d18acb42fe3dee06a730347