Fix flicker near end of quick switch
Unlike isStarted(), mHasLauncherTransitionControllerStarted is true
even after the animation has ended. Once it's ended, we shouldn't
continue updating it even if window shift is still changing.

This can happen when springs are enabled, as that can increase the
recents scroll duration beyond the window/launcher animation
duration (and we updateFinalShift() when the scroll changes).

Test: quick switch by swiping up and over at angle, or really hard
directly to the right, to engage the springs long enough to have a
few frames where you can notice the flicker before this change.

Bug: 147302669
Change-Id: I6a38662612de91352c0f956e6a3137f6c24eba66