Don't scale workspace when going to all apps if ENABLE_OVERVIEW_ACTIONS
Before the two-zone model introduced with ENABLE_OVERVIEW_ACTIONS, we
scaled down the workspace when swiping up, since you could pause at
any time to go to overview (which has workspace scaled down behind it).
But in the two-zone model, that scaling is potentially confusing since
it looks similar to the HintState (when swiping from the nav bar).

Bug: 143361609
Change-Id: I3ee9f6eb45abb0123563d283a3145f27c3a63dd5