Prevent two overlapping long press actions.
We add a check in CheckLongPressHelper to not trigger long press if the
view is already in the pressed state.

This fixes the bug where a stranded DragView is left in the DragLayer.

Repo's every 50 tries:
- Open folder
- Long click and slightly drag the same item over and over in slightly
  different directions.

Logs show that long click on folder item initiates the first long click,
and then later on CheckLongPressHelper triggers another long press before the
initial one is finished.

Bug: 151053858
Change-Id: Ifbb6919e4f0153aaef2904f704baa284e78f1e83