When dragging down on task, base end displacement on thumbnail
TaskView height has a lot of other factors such as icon margin
and actions. Since the thumbnail is what we are dragging to fill
the screen, use its height to determine the progress factor.

Test: swipe down on a task in portrait, landscape, and seascape,
and verify that it follows your finger to the bottom of the
screen. Swiping up also follows your finger for dimiss.

Bug: 151770407
Change-Id: I114a261a06b11f550822178945266cc4a6fd54fe